Character Builder Updates
Update Log 12/15/2015:

Added Random button which picks 35 ranks of perks and calculates a
build.
Added popup mouseover for perk ranks.
Adjusted the priority inputs to hopefully make it more clear that they
can be changed.
Allowed perks to have their own priority.
Allowed the selection of a max rank in the perk.
Changed the sorting on the calculated levels to get a greater variety of
perks earlier. The calculate will now sort every rank of the perks you
have selected by the level and then sort them in the levels by priority.
Changed title in the special box to make it more clear that these are
your intended starting special.
Improved Calculation Speed.