Add perk levels from Far Harbor and Nuka World DLCs
Level requirements and text from http://fallout.wikia.com/wiki/Fallout_4_perks

I prepended the text with "[DLC name]" to make it clear that they're
from those DLCs, in case anyone haven't got them and still want to play
with the tool. Maybe it should be appended instead so it is less
dominant, but this is a start at least.